requirements.txt
# Pipeline Safety AI Evaluator (PSAE)
# Core Dependencies

# Core Framework
numpy>=1.24.0
scipy>=1.10.0

# Statistical Analysis  
pandas>=2.0.0
statsmodels>=0.14.0
pingouin>=0.5.3

# Data Processing
pyyaml>=6.0
jsonschema>=4.17.0

# Progress and UI
tqdm>=4.65.0
colorama>=0.4.6

# AI Model Interfaces (optional - install as needed)
# openai>=1.0.0
# anthropic>=0.8.0
# google-generativeai>=0.3.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0
hypothesis>=6.75.0

# Development
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
